Nexus Tools 4.0

+++ uninstall.sh
@@ -13,54 +13,51 @@
 
 #!/bin/bash
 
-DIR="$HOME/.nexustools"
-ADB="$DIR/adb"
-FASTBOOT="$DIR/fastboot"
-UDEV="/etc/udev/rules.d/51-android.rules"
 OS=$(uname)
 ARCH=$(uname -m)
-XCODE=0
-BASEURL="https://github.com/corbindavenport/nexus-tools/raw/master"
 
-# get sudo
+echo "[INFO] Nexus Tools 4.0"
 
-echo "[INFO] Nexus Tools 3.2"
-echo "[INFO] Please enter sudo password for uninstall."
-sudo echo "[ OK ] Sudo access granted." || { echo "[ERROR] No sudo access."; exit 1; }
-
-# remove files
-
-if [ -f $ADB ]; then
-   sudo rm $ADB
-   echo "[ OK ] ADB removed from $DIR/adb."
-elif [ -f /usr/bin/adb ]; then
-   sudo rm /usr/bin/adb
-   echo "[ OK ] ADB removed from /usr/bin/adb."
-elif [ -f /usr/local/bin/adb ]; then
-   sudo rm /usr/local/bin/adb
-   echo "[ OK ] ADB removed from /usr/local/bin/adb."
-else
-   echo "[INFO] ADB not found in /usr/local/bin or /usr/bin, skipping uninstall."
+# Nexus Tools 3.2+ (August 2016-) installs binaries in ~/.nexustools
+if [ -d $HOME/.nexustools ]; then
+	echo "[WARN] Nexus Tools folder found in $HOME/.nexustools. Press ENTER to delete or X to cancel."
+	read -sn1 input
+	[ "$input" = "" ] && rm -rf $HOME/.nexustools
 fi
 
-if [ -f $FASTBOOT ]; then
-   sudo rm $FASTBOOT
-   echo "[ OK ] Fastboot removed from $DIR/fastboot."
-elif [ -f /usr/bin/fastboot ]; then
-   sudo rm /usr/bin/fastboot
-   echo "[ OK ] ADB removed from /usr/bin/fastboot."
-elif [ -f /usr/local/bin/fastboot ]; then
-   sudo rm /usr/local/bin/fastboot
-   echo "[ OK ] ADB removed from /usr/local/bin/fastboot."
-else
-   echo "[INFO] Fastboot not found in /usr/local/bin or /usr/bin, skipping uninstall."
+# Nexus Tools 2.8-3.1 (October 2015-August 2016) installed binaries in /user/local/bin
+if [ -f /usr/local/bin/adb ]; then
+	echo "[WARN] ADB found in /usr/local/bin. It may have been installed there by an old version of Nexus Tools."
+    echo "[WARN] Sudo access is required to remove. Press ENTER to delete or X to cancel."
+	read -sn1 input
+	[ "$input" = "" ] && sudo rm /usr/local/bin/adb
+fi
+if [ -f /usr/local/bin/fastboot ]; then
+	echo "[WARN] Fasboot found in /usr/local/bin. It may have been installed there by an old version of Nexus Tools."
+    echo "[WARN] Sudo access is required to remove. Press ENTER to delete or X to cancel."
+	read -sn1 input
+	[ "$input" = "" ] && sudo rm /usr/local/bin/fastboot
 fi
 
-if [ -f $UDEV ]; then
-   sudo rm $UDEV
-   echo "[ OK ] Udev list removed."
-else
-   echo "[INFO] Udev list not found in /etc/udev/rules.d/, skipping uninstall."
+# Nexus Tools 1.0-2.7.1 (December 2013-October 2015) installed binaries in /usr/bin
+if [ -f /usr/bin/adb ]; then
+	echo "[WARN] ADB found in /usr/bin. It may have been installed there by an old version of Nexus Tools."
+    echo "[WARN] Sudo access is required to remove. Press ENTER to delete or X to cancel."
+	read -sn1 input
+	[ "$input" = "" ] && sudo rm /usr/bin/adb
 fi
-echo "[ OK ] Done uninstalling."
-echo " "
+if [ -f /usr/bin/fastboot ]; then
+	echo "[WARN] Fasboot found in /usr/bin. It may have been installed there by an old version of Nexus Tools."
+    echo "[WARN] Sudo access is required to remove. Press ENTER to delete or X to cancel."
+	read -sn1 input
+	[ "$input" = "" ] && sudo rm /usr/bin/fastboot
+fi
+
+# Uninstall Device ID list
+if [ -f $HOME/.android/adb_usb.ini ]; then
+	rm $HOME/.android/adb_usb.ini
+	echo "[ OK ] Deleted $HOME/.android/adb_usb.ini"
+fi
+
+echo "[ OK ] Uninstall complete."
+echo ""
